# src/docs/swagger.yaml
basePath: /
definitions:
...
  routes.CreateRoleRequest:
    properties:
      assign_to_existing:
        type: boolean
      custom_id:
        type: string
      description:
        type: string
      is_default:
        type: boolean
      name:
        type: string
      permission_bitmask:
        minimum: 0
        type: integer
    required:
    - custom_id
    - name
    type: object
...
  routes.CreateUserRoleRequest:
    properties:
      role_id:
        type: string
    required:
    - role_id
    type: object
...
  routes.ProfileDTO:
    properties:
      bio:
        type: string
      birthdate:
        type: string
      birthdate_visible:
        type: boolean
      display_name:
        type: string
      joined_at:
        type: string
      twitter_handle:
        type: string
      user_id:
        type: string
      website_url:
        type: string
    type: object
...
  routes.RoleDTO:
    properties:
      custom_id:
        type: string
      description:
        type: string
      id:
        type: string
      is_default:
        type: boolean
      name:
        type: string
      permission_bitmask:
        type: integer
    type: object
  routes.RoleListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/routes.RoleDTO'
        type: array
    type: object
...
  routes.UpdateRoleRequest:
    properties:
      description:
        type: string
      is_default:
        type: boolean
      name:
        type: string
      permission_bitmask:
        type: integer
    type: object
...
  routes.UserDTO:
    properties:
      affiliation_period:
        type: string
      created_at:
        type: string
      custom_id:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      external_email:
        type: string
      id:
        type: string
      pending_email:
        type: string
      profile:
        $ref: '#/definitions/routes.ProfileDTO'
      status:
        type: string
      updated_at:
        type: string
    type: object
  routes.UserListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/routes.UserDTO'
        type: array
    type: object
info:
  contact: {}
paths:
...
  /roles:
    get:
      description: List roles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.RoleListResponse'
      summary: List roles
      tags:
      - roles
    post:
      consumes:
      - application/json
      description: Create a new role
      parameters:
      - description: Create role
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/routes.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/routes.RoleDTO'
      summary: Create a role
      tags:
      - roles
  /roles/{id}:
    get:
      description: Get a single role
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.RoleDTO'
      summary: Get a role
      tags:
      - roles
    put:
      consumes:
      - application/json
      description: Update role fields
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Update role
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/routes.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.RoleDTO'
      summary: Update a role
      tags:
      - roles
  /roles/{id}/assign_all:
    post:
      description: Assign the role to all users with status active or established
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Assign role to all existing users
      tags:
      - roles
  /roles/{id}/users:
    get:
      description: Get users assigned to a role
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.UserListResponse'
      summary: List users for a role
      tags:
      - roles
...
  /users/{id}/roles:
    get:
      description: Get roles assigned to a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.RoleListResponse'
      summary: List roles for a user
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Assign the specified role to the user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Role assignment
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/routes.CreateUserRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/routes.RoleDTO'
      summary: Assign a role to a user
      tags:
      - users
  /users/{id}/roles/{roleId}:
    delete:
      description: Unassign the specified role from the user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Remove a role from a user
      tags:
      - users
swagger: "2.0"